format conv fully
