--- kernels/conv/src/conv2d_3x3rvv.cpp
@@ -1,712 +1,0 @@
-#include <riscv_vector.h>
-#include <stddef.h>
-#include <stdint.h>
-#include <string.h>
-#include <stdlib.h>
-
-// ============================================================================
-// HELPER FUNCTIONS
-// ============================================================================
-
-// Create zero-padded input (pad=1 for 3x3 kernel)
-float* create_padded_input(const float* input, int H, int W) {
-    int H_pad = H + 2;
-    int W_pad = W + 2;
-    
-    float* padded = (float*)calloc(H_pad * W_pad, sizeof(float));
-    
-    // Copy input to center of padded buffer
-    for (int h = 0; h < H; h++) {
-        memcpy(padded + (h + 1) * W_pad + 1, 
-               input + h * W, 
-               W * sizeof(float));
-    }
-    
-    return padded;
-}
-
-// ============================================================================
-// M1 IMPLEMENTATION (VLEN=128 → 4 float32 elements per iteration)
-// ============================================================================
-
-void conv3x3_rvv_m1(
-    const float* input,    // Input: HxW
-    const float* kernel,   // Kernel: 3x3 (9 elements, row-major)
-    float* output,         // Output: HxW (with padding)
-    int H,                 // Input height
-    int W,                 // Input width
-    bool use_padding       // If true, applies zero-padding
-) {
-    // Create padded input if needed
-    float* padded_input = nullptr;
-    const float* proc_input = input;
-    int H_proc = H;
-    int W_proc = W;
-    
-    if (use_padding) {
-        padded_input = create_padded_input(input, H, W);
-        proc_input = padded_input;
-        H_proc = H + 2;
-        W_proc = W + 2;
-    }
-    
-    const int out_h = use_padding ? H : (H - 2);
-    const int out_w = use_padding ? W : (W - 2);
-    
-    // Load kernel weights into scalar registers for broadcasting
-    float k00 = kernel[0], k01 = kernel[1], k02 = kernel[2];
-    float k10 = kernel[3], k11 = kernel[4], k12 = kernel[5];
-    float k20 = kernel[6], k21 = kernel[7], k22 = kernel[8];
-    
-    // Process each output row
-    for (int oh = 0; oh < out_h; oh++) {
-        // Get pointers to the 3 input rows needed for this output row
-        const float* row0 = proc_input + oh * W_proc;
-        const float* row1 = row0 + W_proc;
-        const float* row2 = row1 + W_proc;
-        float* out_row = output + oh * out_w;
-        
-        // Vectorized processing across output width with m1
-        int ow = 0;
-        while (ow < out_w) {
-            // Set vector length (m1: LMUL=1, processes ~4 elements on VLEN=128)
-            size_t vl = __riscv_vsetvl_e32m1(out_w - ow);
-            
-            // ================================================================
-            // LOAD PHASE: Load 9 vectors (3x3 neighborhood)
-            // ================================================================
-            // Row 0: Load 3 overlapping vectors with offsets 0, 1, 2
-            vfloat32m1_t v00 = __riscv_vle32_v_f32m1(row0 + ow, vl);
-            vfloat32m1_t v01 = __riscv_vle32_v_f32m1(row0 + ow + 1, vl);
-            vfloat32m1_t v02 = __riscv_vle32_v_f32m1(row0 + ow + 2, vl);
-            
-            // Row 1: Load 3 overlapping vectors
-            vfloat32m1_t v10 = __riscv_vle32_v_f32m1(row1 + ow, vl);
-            vfloat32m1_t v11 = __riscv_vle32_v_f32m1(row1 + ow + 1, vl);
-            vfloat32m1_t v12 = __riscv_vle32_v_f32m1(row1 + ow + 2, vl);
-            
-            // Row 2: Load 3 overlapping vectors
-            vfloat32m1_t v20 = __riscv_vle32_v_f32m1(row2 + ow, vl);
-            vfloat32m1_t v21 = __riscv_vle32_v_f32m1(row2 + ow + 1, vl);
-            vfloat32m1_t v22 = __riscv_vle32_v_f32m1(row2 + ow + 2, vl);
-            
-            // ================================================================
-            // COMPUTE PHASE: Fused Multiply-Accumulate (FMA) chain
-            // ================================================================
-            // Start with first multiply
-            vfloat32m1_t acc = __riscv_vfmul_vf_f32m1(v00, k00, vl);
-            
-            // Chain the remaining 8 FMAs (vector * scalar + accumulator)
-            acc = __riscv_vfmacc_vf_f32m1(acc, k01, v01, vl);
-            acc = __riscv_vfmacc_vf_f32m1(acc, k02, v02, vl);
-            acc = __riscv_vfmacc_vf_f32m1(acc, k10, v10, vl);
-            acc = __riscv_vfmacc_vf_f32m1(acc, k11, v11, vl);
-            acc = __riscv_vfmacc_vf_f32m1(acc, k12, v12, vl);
-            acc = __riscv_vfmacc_vf_f32m1(acc, k20, v20, vl);
-            acc = __riscv_vfmacc_vf_f32m1(acc, k21, v21, vl);
-            acc = __riscv_vfmacc_vf_f32m1(acc, k22, v22, vl);
-            
-            // ================================================================
-            // STORE PHASE: Write results
-            // ================================================================
-            __riscv_vse32_v_f32m1(out_row + ow, acc, vl);
-            
-            // Move to next vector chunk
-            ow += vl;
-        }
-    }
-    
-    // Cleanup
-    if (padded_input) {
-        free(padded_input);
-    }
-}
-
-// ============================================================================
-// M2 IMPLEMENTATION (VLEN=128 → 8 float32 elements per iteration)
-// ============================================================================
-
-void conv3x3_rvv_m2(
-    const float* input,    // Input: HxW
-    const float* kernel,   // Kernel: 3x3 (9 elements, row-major)
-    float* output,         // Output: HxW (with padding) or (H-2)x(W-2)
-    int H,                 // Input height
-    int W,                 // Input width
-    bool use_padding       // If true, applies zero-padding
-) {
-    // Create padded input if needed
-    float* padded_input = nullptr;
-    const float* proc_input = input;
-    int H_proc = H;
-    int W_proc = W;
-    
-    if (use_padding) {
-        padded_input = create_padded_input(input, H, W);
-        proc_input = padded_input;
-        H_proc = H + 2;
-        W_proc = W + 2;
-    }
-    
-    const int out_h = use_padding ? H : (H - 2);
-    const int out_w = use_padding ? W : (W - 2);
-    
-    // Load kernel weights into scalar registers
-    float k00 = kernel[0], k01 = kernel[1], k02 = kernel[2];
-    float k10 = kernel[3], k11 = kernel[4], k12 = kernel[5];
-    float k20 = kernel[6], k21 = kernel[7], k22 = kernel[8];
-    
-    // Process each output row
-    for (int oh = 0; oh < out_h; oh++) {
-        // Get pointers to the 3 input rows needed
-        const float* row0 = proc_input + oh * W_proc;
-        const float* row1 = row0 + W_proc;
-        const float* row2 = row1 + W_proc;
-        float* out_row = output + oh * out_w;
-        
-        // Vectorized processing with m2 (2x throughput vs m1)
-        int ow = 0;
-        while (ow < out_w) {
-            // Set vector length (m2: LMUL=2, processes ~8 elements on VLEN=128)
-            size_t vl = __riscv_vsetvl_e32m2(out_w - ow);
-            
-            // ================================================================
-            // LOAD PHASE: Load 9 vectors with m2
-            // ================================================================
-            // Row 0: 3 vectors with offsets 0, 1, 2
-            vfloat32m2_t v00 = __riscv_vle32_v_f32m2(row0 + ow, vl);
-            vfloat32m2_t v01 = __riscv_vle32_v_f32m2(row0 + ow + 1, vl);
-            vfloat32m2_t v02 = __riscv_vle32_v_f32m2(row0 + ow + 2, vl);
-            
-            // Row 1: 3 vectors
-            vfloat32m2_t v10 = __riscv_vle32_v_f32m2(row1 + ow, vl);
-            vfloat32m2_t v11 = __riscv_vle32_v_f32m2(row1 + ow + 1, vl);
-            vfloat32m2_t v12 = __riscv_vle32_v_f32m2(row1 + ow + 2, vl);
-            
-            // Row 2: 3 vectors
-            vfloat32m2_t v20 = __riscv_vle32_v_f32m2(row2 + ow, vl);
-            vfloat32m2_t v21 = __riscv_vle32_v_f32m2(row2 + ow + 1, vl);
-            vfloat32m2_t v22 = __riscv_vle32_v_f32m2(row2 + ow + 2, vl);
-            
-            // ================================================================
-            // COMPUTE PHASE: FMA chain with m2 vectors
-            // ================================================================
-            vfloat32m2_t acc = __riscv_vfmul_vf_f32m2(v00, k00, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k01, v01, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k02, v02, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k10, v10, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k11, v11, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k12, v12, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k20, v20, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k21, v21, vl);
-            acc = __riscv_vfmacc_vf_f32m2(acc, k22, v22, vl);
-            
-            // ================================================================
-            // STORE PHASE
-            // ================================================================
-            __riscv_vse32_v_f32m2(out_row + ow, acc, vl);
-            
-            ow += vl;
-        }
-    }
-    
-    // Cleanup
-    if (padded_input) {
-        free(padded_input);
-    }
-}
-
-// ============================================================================
-// MULTI-ROW BATCHED M2 (Cache-optimized for edge devices)
-// ============================================================================
-
-void conv3x3_rvv_m2_batched(
-    const float* input,
-    const float* kernel,
-    float* output,
-    int H,
-    int W,
-    bool use_padding,
-    int batch_rows = 4    // Process N output rows together
-) {
-    float* padded_input = nullptr;
-    const float* proc_input = input;
-    int H_proc = H;
-    int W_proc = W;
-    
-    if (use_padding) {
-        padded_input = create_padded_input(input, H, W);
-        proc_input = padded_input;
-        H_proc = H + 2;
-        W_proc = W + 2;
-    }
-    
-    const int out_h = use_padding ? H : (H - 2);
-    const int out_w = use_padding ? W : (W - 2);
-    
-    float k00 = kernel[0], k01 = kernel[1], k02 = kernel[2];
-    float k10 = kernel[3], k11 = kernel[4], k12 = kernel[5];
-    float k20 = kernel[6], k21 = kernel[7], k22 = kernel[8];
-    
-    // Process output rows in batches for better cache reuse
-    for (int oh_base = 0; oh_base < out_h; oh_base += batch_rows) {
-        int rows_to_process = (oh_base + batch_rows <= out_h) ? 
-                              batch_rows : (out_h - oh_base);
-        
-        // For each column position (vectorized)
-        for (int ow = 0; ow < out_w; ) {
-            size_t vl = __riscv_vsetvl_e32m2(out_w - ow);
-            
-            // Process each row in the current batch
-            for (int r = 0; r < rows_to_process; r++) {
-                int oh = oh_base + r;
-                const float* row0 = proc_input + oh * W_proc;
-                const float* row1 = row0 + W_proc;
-                const float* row2 = row1 + W_proc;
-                
-                // Load 9 vectors
-                vfloat32m2_t v00 = __riscv_vle32_v_f32m2(row0 + ow, vl);
-                vfloat32m2_t v01 = __riscv_vle32_v_f32m2(row0 + ow + 1, vl);
-                vfloat32m2_t v02 = __riscv_vle32_v_f32m2(row0 + ow + 2, vl);
-                
-                vfloat32m2_t v10 = __riscv_vle32_v_f32m2(row1 + ow, vl);
-                vfloat32m2_t v11 = __riscv_vle32_v_f32m2(row1 + ow + 1, vl);
-                vfloat32m2_t v12 = __riscv_vle32_v_f32m2(row1 + ow + 2, vl);
-                
-                vfloat32m2_t v20 = __riscv_vle32_v_f32m2(row2 + ow, vl);
-                vfloat32m2_t v21 = __riscv_vle32_v_f32m2(row2 + ow + 1, vl);
-                vfloat32m2_t v22 = __riscv_vle32_v_f32m2(row2 + ow + 2, vl);
-                
-                // Compute
-                vfloat32m2_t acc = __riscv_vfmul_vf_f32m2(v00, k00, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k01, v01, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k02, v02, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k10, v10, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k11, v11, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k12, v12, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k20, v20, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k21, v21, vl);
-                acc = __riscv_vfmacc_vf_f32m2(acc, k22, v22, vl);
-                
-                // Store
-                __riscv_vse32_v_f32m2(output + oh * out_w + ow, acc, vl);
-            }
-            
-            ow += vl;
-        }
-    }
-    
-    if (padded_input) {
-        free(padded_input);
-    }
-}
-
-// ============================================================================
-// RGB / 3-CHANNEL SUPPORT
-// ============================================================================
-
-void conv3x3_rvv_m2_rgb(
-    const float* input,     // Input: 3xHxW (channel-major layout)
-    const float* kernel,    // Kernel: 3x3x3 (27 elements: R, G, B kernels)
-    float* output,          // Output: 3xHxW
-    int H,
-    int W,
-    bool use_padding
-) {
-    const int out_h = use_padding ? H : (H - 2);
-    const int out_w = use_padding ? W : (W - 2);
-    
-    // Process each channel independently
-    for (int c = 0; c < 3; c++) {
-        const float* in_channel = input + c * H * W;
-        const float* kernel_channel = kernel + c * 9;
-        float* out_channel = output + c * out_h * out_w;
-        
-        conv3x3_rvv_m2(in_channel, kernel_channel, out_channel, H, W, use_padding);
-    }
-}
-
-// ============================================================================
-// EXAMPLE USAGE AND TESTING
-// ============================================================================
-
-#ifdef EXAMPLE_USAGE
-#include <stdio.h>
-#include <time.h>
-#include <math.h>
-
-// Include general conv2d implementations for comparison
-// These are C++ functions from conv2d_rvv.cpp and conv2d_scalar.cpp
-void conv2d_scalar(
-    const float* input, const float* kernel, float* output,
-    int batch_size, int in_channels, int out_channels,
-    int input_h, int input_w, int kernel_h, int kernel_w,
-    int stride_h, int stride_w, int pad_h, int pad_w);
-
-void conv2d_e32m1(
-    const float* input, const float* kernel, float* output,
-    int batch_size, int in_channels, int out_channels,
-    int input_h, int input_w, int kernel_h, int kernel_w,
-    int stride_h, int stride_w, int pad_h, int pad_w);
-
-void conv2d_e32m2(
-    const float* input, const float* kernel, float* output,
-    int batch_size, int in_channels, int out_channels,
-    int input_h, int input_w, int kernel_h, int kernel_w,
-    int stride_h, int stride_w, int pad_h, int pad_w);
-
-void print_matrix(const float* mat, int H, int W, const char* name) {
-    printf("\n%s (%dx%d):\n", name, H, W);
-    for (int i = 0; i < H; i++) {
-        for (int j = 0; j < W; j++) {
-            printf("%6.2f ", mat[i * W + j]);
-        }
-        printf("\n");
-    }
-}
-
-// Simple timing function (returns time in seconds)
-double get_time() {
-    struct timespec ts;
-    clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec + ts.tv_nsec * 1e-9;
-}
-
-// Compare two output matrices
-bool compare_outputs(const float* a, const float* b, int size, float epsilon = 1e-5) {
-    for (int i = 0; i < size; i++) {
-        if (fabs(a[i] - b[i]) > epsilon) {
-            printf("Mismatch at index %d: %.6f vs %.6f\n", i, a[i], b[i]);
-            return false;
-        }
-    }
-    return true;
-}
-
-int main() {
-    printf("========================================\n");
-    printf("RVV 3x3 Convolution Test Suite\n");
-    printf("========================================\n");
-    
-    // ========================================================================
-    // TEST 1: Small 5x5 image with averaging kernel
-    // ========================================================================
-    printf("\n[TEST 1] 5x5 Image with Averaging Kernel\n");
-    printf("------------------------------------------\n");
-    
-    const int H = 5, W = 5;
-    float input[H * W];
-    
-    // Create a simple test pattern
-    for (int i = 0; i < H; i++) {
-        for (int j = 0; j < W; j++) {
-            input[i * W + j] = (float)(i * W + j + 1);
-        }
-    }
-    
-    // 3x3 Averaging kernel (box blur)
-    float kernel[9] = {
-        1.0f/9.0f, 1.0f/9.0f, 1.0f/9.0f,
-        1.0f/9.0f, 1.0f/9.0f, 1.0f/9.0f,
-        1.0f/9.0f, 1.0f/9.0f, 1.0f/9.0f
-    };
-    
-    // Output buffers
-    float output_m1[H * W];
-    float output_m2[H * W];
-    float output_batched[H * W];
-    
-    // Test m1
-    conv3x3_rvv_m1(input, kernel, output_m1, H, W, true);
-    print_matrix(input, H, W, "Input");
-    print_matrix(output_m1, H, W, "Output (m1, with padding)");
-    
-    // Test m2
-    conv3x3_rvv_m2(input, kernel, output_m2, H, W, true);
-    print_matrix(output_m2, H, W, "Output (m2, with padding)");
-    
-    // Verify m1 and m2 match
-    bool match = compare_outputs(output_m1, output_m2, H * W);
-    printf("\n✓ m1 and m2 results: %s\n", match ? "MATCH" : "DIFFER");
-    
-    // ========================================================================
-    // TEST 2: Batched versions with different batch sizes
-    // ========================================================================
-    printf("\n[TEST 2] Batched M2 with Different Batch Sizes\n");
-    printf("------------------------------------------------\n");
-    
-    int batch_sizes[] = {1, 2, 4};
-    int num_batch_sizes = sizeof(batch_sizes) / sizeof(batch_sizes[0]);
-    
-    for (int i = 0; i < num_batch_sizes; i++) {
-        int batch = batch_sizes[i];
-        printf("\nTesting batch_rows = %d...\n", batch);
-        
-        conv3x3_rvv_m2_batched(input, kernel, output_batched, H, W, true, batch);
-        
-        bool batch_match = compare_outputs(output_m2, output_batched, H * W);
-        printf("  Result vs m2: %s\n", batch_match ? "MATCH ✓" : "DIFFER ✗");
-        
-        if (!batch_match) {
-            print_matrix(output_batched, H, W, "Batched Output (mismatch)");
-        }
-    }
-    
-    // ========================================================================
-    // TEST 3: Performance Benchmark
-    // ========================================================================
-    printf("\n[TEST 3] Performance Benchmark\n");
-    printf("--------------------------------\n");
-    
-    // Benchmark parameters
-    const int BENCH_H = 256, BENCH_W = 256;
-    const int WARMUP_ITERS = 10;
-    const int BENCH_ITERS = 100;
-    
-    // Allocate benchmark buffers
-    float* bench_input = (float*)malloc(BENCH_H * BENCH_W * sizeof(float));
-    float* bench_output = (float*)malloc(BENCH_H * BENCH_W * sizeof(float));
-    
-    // Initialize with random-ish data
-    for (int i = 0; i < BENCH_H * BENCH_W; i++) {
-        bench_input[i] = (float)(i % 255);
-    }
-    
-    printf("\nBenchmarking on %dx%d image (%d iterations)...\n", 
-           BENCH_H, BENCH_W, BENCH_ITERS);
-    
-    // ---- Benchmark M1 ----
-    printf("\n[M1 Implementation]\n");
-    
-    // Warmup
-    for (int i = 0; i < WARMUP_ITERS; i++) {
-        conv3x3_rvv_m1(bench_input, kernel, bench_output, BENCH_H, BENCH_W, true);
-    }
-    
-    // Timed run
-    double start = get_time();
-    for (int i = 0; i < BENCH_ITERS; i++) {
-        conv3x3_rvv_m1(bench_input, kernel, bench_output, BENCH_H, BENCH_W, true);
-    }
-    double end = get_time();
-    
-    double m1_time = (end - start) / BENCH_ITERS;
-    double m1_mpixels_per_sec = (BENCH_H * BENCH_W / 1e6) / m1_time;
-    printf("  Average time: %.6f seconds\n", m1_time);
-    printf("  Throughput:   %.2f MPixels/sec\n", m1_mpixels_per_sec);
-    
-    // ---- Benchmark M2 ----
-    printf("\n[M2 Implementation]\n");
-    
-    // Warmup
-    for (int i = 0; i < WARMUP_ITERS; i++) {
-        conv3x3_rvv_m2(bench_input, kernel, bench_output, BENCH_H, BENCH_W, true);
-    }
-    
-    // Timed run
-    start = get_time();
-    for (int i = 0; i < BENCH_ITERS; i++) {
-        conv3x3_rvv_m2(bench_input, kernel, bench_output, BENCH_H, BENCH_W, true);
-    }
-    end = get_time();
-    
-    double m2_time = (end - start) / BENCH_ITERS;
-    double m2_mpixels_per_sec = (BENCH_H * BENCH_W / 1e6) / m2_time;
-    printf("  Average time: %.6f seconds\n", m2_time);
-    printf("  Throughput:   %.2f MPixels/sec\n", m2_mpixels_per_sec);
-    printf("  Speedup vs M1: %.2fx\n", m1_time / m2_time);
-    
-    // ---- Benchmark M2 Batched (different batch sizes) ----
-    printf("\n[M2 Batched Implementation]\n");
-    
-    int bench_batches[] = {2, 4, 8};
-    int num_bench_batches = sizeof(bench_batches) / sizeof(bench_batches[0]);
-    
-    for (int i = 0; i < num_bench_batches; i++) {
-        int batch = bench_batches[i];
-        
-        // Warmup
-        for (int j = 0; j < WARMUP_ITERS; j++) {
-            conv3x3_rvv_m2_batched(bench_input, kernel, bench_output, 
-                                   BENCH_H, BENCH_W, true, batch);
-        }
-        
-        // Timed run
-        start = get_time();
-        for (int j = 0; j < BENCH_ITERS; j++) {
-            conv3x3_rvv_m2_batched(bench_input, kernel, bench_output, 
-                                   BENCH_H, BENCH_W, true, batch);
-        }
-        end = get_time();
-        
-        double batched_time = (end - start) / BENCH_ITERS;
-        double batched_mpixels = (BENCH_H * BENCH_W / 1e6) / batched_time;
-        
-        printf("  batch_rows=%d:\n", batch);
-        printf("    Average time: %.6f seconds\n", batched_time);
-        printf("    Throughput:   %.2f MPixels/sec\n", batched_mpixels);
-        printf("    Speedup vs M1: %.2fx\n", m1_time / batched_time);
-        printf("    Speedup vs M2: %.2fx\n", m2_time / batched_time);
-    }
-    
-    // ========================================================================
-    // TEST 4: Compare against General RVV Implementation
-    // ========================================================================
-    printf("\n[TEST 4] Comparison with General RVV Implementation\n");
-    printf("----------------------------------------------------\n");
-    printf("Comparing specialized 3x3 vs general conv2d (NCHW format)\n\n");
-    
-    // Prepare data in NCHW format for general implementation
-    const int batch = 1;
-    const int in_ch = 1;
-    const int out_ch = 1;
-    
-    // Flatten kernel for NCHW: [out_ch, in_ch, kH, kW]
-    float general_kernel[1 * 1 * 3 * 3];
-    for (int i = 0; i < 9; i++) {
-        general_kernel[i] = kernel[i];
-    }
-    
-    // Allocate output buffer for general implementation
-    float* general_output = (float*)malloc(BENCH_H * BENCH_W * sizeof(float));
-    
-    // ---- Correctness Check: General Scalar ----
-    printf("[Correctness] Testing general scalar implementation...\n");
-    conv2d_scalar(bench_input, general_kernel, general_output,
-                  batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    
-    bool scalar_match = compare_outputs(bench_output, general_output, BENCH_H * BENCH_W);
-    printf("  Specialized 3x3 m2 vs General scalar: %s\n", 
-           scalar_match ? "MATCH ✓" : "DIFFER ✗");
-    
-    // ---- Correctness Check: General RVV e32m1 ----
-    printf("[Correctness] Testing general RVV e32m1 implementation...\n");
-    conv2d_e32m1(bench_input, general_kernel, general_output,
-                 batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    
-    bool e32m1_match = compare_outputs(bench_output, general_output, BENCH_H * BENCH_W);
-    printf("  Specialized 3x3 m2 vs General e32m1: %s\n", 
-           e32m1_match ? "MATCH ✓" : "DIFFER ✗");
-    
-    // ---- Correctness Check: General RVV e32m2 ----
-    printf("[Correctness] Testing general RVV e32m2 implementation...\n");
-    conv2d_e32m2(bench_input, general_kernel, general_output,
-                 batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    
-    bool e32m2_match = compare_outputs(bench_output, general_output, BENCH_H * BENCH_W);
-    printf("  Specialized 3x3 m2 vs General e32m2: %s\n\n", 
-           e32m2_match ? "MATCH ✓" : "DIFFER ✗");
-    
-    // ---- Performance Benchmark: General Scalar ----
-    printf("[Performance] Benchmarking general scalar implementation...\n");
-    
-    // Warmup
-    for (int i = 0; i < WARMUP_ITERS; i++) {
-        conv2d_scalar(bench_input, general_kernel, general_output,
-                      batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    }
-    
-    // Timed run
-    start = get_time();
-    for (int i = 0; i < BENCH_ITERS; i++) {
-        conv2d_scalar(bench_input, general_kernel, general_output,
-                      batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    }
-    end = get_time();
-    
-    double general_scalar_time = (end - start) / BENCH_ITERS;
-    double general_scalar_mpixels = (BENCH_H * BENCH_W / 1e6) / general_scalar_time;
-    printf("  Average time: %.6f seconds\n", general_scalar_time);
-    printf("  Throughput:   %.2f MPixels/sec\n", general_scalar_mpixels);
-    printf("  Speedup (3x3 m2 vs general scalar): %.2fx\n\n", 
-           general_scalar_time / m2_time);
-    
-    // ---- Performance Benchmark: General RVV e32m1 ----
-    printf("[Performance] Benchmarking general RVV e32m1 implementation...\n");
-    
-    // Warmup
-    for (int i = 0; i < WARMUP_ITERS; i++) {
-        conv2d_e32m1(bench_input, general_kernel, general_output,
-                     batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    }
-    
-    // Timed run
-    start = get_time();
-    for (int i = 0; i < BENCH_ITERS; i++) {
-        conv2d_e32m1(bench_input, general_kernel, general_output,
-                     batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    }
-    end = get_time();
-    
-    double general_e32m1_time = (end - start) / BENCH_ITERS;
-    double general_e32m1_mpixels = (BENCH_H * BENCH_W / 1e6) / general_e32m1_time;
-    printf("  Average time: %.6f seconds\n", general_e32m1_time);
-    printf("  Throughput:   %.2f MPixels/sec\n", general_e32m1_mpixels);
-    printf("  Speedup (3x3 m1 vs general e32m1): %.2fx\n", 
-           general_e32m1_time / m1_time);
-    printf("  Speedup (3x3 m2 vs general e32m1): %.2fx\n\n", 
-           general_e32m1_time / m2_time);
-    
-    // ---- Performance Benchmark: General RVV e32m2 ----
-    printf("[Performance] Benchmarking general RVV e32m2 implementation...\n");
-    
-    // Warmup
-    for (int i = 0; i < WARMUP_ITERS; i++) {
-        conv2d_e32m2(bench_input, general_kernel, general_output,
-                     batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    }
-    
-    // Timed run
-    start = get_time();
-    for (int i = 0; i < BENCH_ITERS; i++) {
-        conv2d_e32m2(bench_input, general_kernel, general_output,
-                     batch, in_ch, out_ch, BENCH_H, BENCH_W, 3, 3, 1, 1, 1, 1);
-    }
-    end = get_time();
-    
-    double general_e32m2_time = (end - start) / BENCH_ITERS;
-    double general_e32m2_mpixels = (BENCH_H * BENCH_W / 1e6) / general_e32m2_time;
-    printf("  Average time: %.6f seconds\n", general_e32m2_time);
-    printf("  Throughput:   %.2f MPixels/sec\n", general_e32m2_mpixels);
-    printf("  Speedup (3x3 m1 vs general e32m2): %.2fx\n", 
-           general_e32m2_time / m1_time);
-    printf("  Speedup (3x3 m2 vs general e32m2): %.2fx\n\n", 
-           general_e32m2_time / m2_time);
-    
-    // Summary table
-    printf("========================================\n");
-    printf("Performance Comparison Table\n");
-    printf("========================================\n");
-    printf("%-30s %12s %15s\n", "Implementation", "Time (sec)", "Throughput");
-    printf("------------------------------------------------------------\n");
-    printf("%-30s %12.6f %12.2f MP/s\n", "General Scalar", general_scalar_time, general_scalar_mpixels);
-    printf("%-30s %12.6f %12.2f MP/s\n", "General RVV e32m1", general_e32m1_time, general_e32m1_mpixels);
-    printf("%-30s %12.6f %12.2f MP/s\n", "General RVV e32m2", general_e32m2_time, general_e32m2_mpixels);
-    printf("%-30s %12.6f %12.2f MP/s\n", "Specialized 3x3 M1", m1_time, m1_mpixels_per_sec);
-    printf("%-30s %12.6f %12.2f MP/s  ← FASTEST\n", "Specialized 3x3 M2", m2_time, m2_mpixels_per_sec);
-    printf("------------------------------------------------------------\n");
-    printf("Overall Speedup (3x3 M2 vs General e32m2): %.2fx\n", 
-           general_e32m2_time / m2_time);
-    
-    free(general_output);
-    
-    // ========================================================================
-    // Summary
-    // ========================================================================
-    printf("\n========================================\n");
-    printf("Summary\n");
-    printf("========================================\n");
-    printf("All tests completed successfully!\n");
-    printf("  ✓ Correctness: All implementations produce identical results\n");
-    printf("  ✓ Specialized 3x3: ~%.1fx faster than general RVV e32m2\n", 
-           general_e32m2_time / m2_time);
-    printf("  ✓ Specialized 3x3: ~%.1fx faster than general scalar\n", 
-           general_scalar_time / m2_time);
-    printf("  ✓ Batching: Tested with batch sizes 1, 2, 4, 8\n");
-    
-    // Cleanup
-    free(bench_input);
-    free(bench_output);
-    
-    return 0;
-}
-#endif
