--- kernels/conv/src/conv2d_rvv.cpp
@@ -1,446 +1,0 @@
-#include <riscv_vector.h>
-#include <cstring>
-#include <algorithm>
-#include <iostream>
-#include <cmath>
-#include "defs.h"
-#include "rvv_defs.hpp"
-
-// RVV optimized 2D convolution (e32m1)
-void conv2d_e32m1(
-    const float* input, const float* kernel, float* output,
-    int batch_size, int in_channels, int out_channels,
-    int input_h, int input_w, int kernel_h, int kernel_w,
-    int stride_h, int stride_w, int pad_h, int pad_w) {
-    
-    // Calculate output dimensions
-    int out_height = (input_h + 2 * pad_h - kernel_h) / stride_h + 1;
-    int out_width = (input_w + 2 * pad_w - kernel_w) / stride_w + 1;
-    
-    // Initialize output to zero
-    size_t output_size = batch_size * out_channels * out_height * out_width;
-    std::memset(output, 0, output_size * sizeof(float));
-    
-    for (int b = 0; b < batch_size; ++b) {
-        for (int oc = 0; oc < out_channels; ++oc) {
-            for (int oh = 0; oh < out_height; ++oh) {
-                for (int ow = 0; ow < out_width; ++ow) {
-                    
-                    // Calculate input region bounds
-                    int in_h_start = oh * stride_h - pad_h;
-                    int in_w_start = ow * stride_w - pad_w;
-                    int in_h_end = in_h_start + kernel_h;
-                    int in_w_end = in_w_start + kernel_w;
-                    
-                    float sum = 0.0f;
-                    
-                    for (int ic = 0; ic < in_channels; ++ic) {
-                        for (int kh = 0; kh < kernel_h; ++kh) {
-                            int in_h = in_h_start + kh;
-                            
-                            // Skip if outside input bounds
-                            if (in_h < 0 || in_h >= input_h) {
-                                continue;
-                            }
-                            
-                            int kw = 0;
-                            int in_w = in_w_start;
-                            
-                            // Skip negative width indices
-                            while (kw < kernel_w && in_w + kw < 0) {
-                                kw++;
-                            }
-                            
-                            size_t vl;
-                            for (; kw < kernel_w; kw += vl) {
-                                int remaining = kernel_w - kw;
-                                int valid_end = input_w - in_w - kw;
-                                int processable = std::min(remaining, valid_end);
-                                
-                                if (processable <= 0) break;
-                                
-                                vl = SET_VECTOR_LENGTH<float, M1>(processable);
-
-                                // Load input and kernel vectors using wrappers
-                                vfloat32m1_t v_input = VECTOR_LOAD<float, M1>(
-                                    &input[b * in_channels * input_h * input_w +
-                                           ic * input_h * input_w +
-                                           in_h * input_w + in_w + kw], vl);
-
-                                vfloat32m1_t v_kernel = VECTOR_LOAD<float, M1>(
-                                    &kernel[oc * in_channels * kernel_h * kernel_w +
-                                           ic * kernel_h * kernel_w +
-                                           kh * kernel_w + kw], vl);
-
-                                // Element-wise multiply
-                                vfloat32m1_t v_mult = VECTOR_MUL_VV<float, M1>(v_input, v_kernel, vl);
-
-                                // Reduce sum horizontally
-                                vfloat32m1_t v_zero = VECTOR_BROADCAST<float, M1>(0.0f, 1);
-                                vfloat32m1_t v_sum = __riscv_vfredusum_vs_f32m1_f32m1(v_mult, v_zero, vl);
-
-                                // Extract scalar sum and add to total
-                                sum += __riscv_vfmv_f_s_f32m1_f32(v_sum);
-                            }
-                        }
-                    }
-                    
-                    // Store the accumulated result
-                    output[b * out_channels * out_height * out_width +
-                           oc * out_height * out_width +
-                           oh * out_width + ow] = sum;
-                }
-            }
-        }
-    }
-}
-
-// RVV optimized 2D convolution (e32m2)
-void conv2d_e32m2(
-    const float* input, const float* kernel, float* output,
-    int batch_size, int in_channels, int out_channels,
-    int input_h, int input_w, int kernel_h, int kernel_w,
-    int stride_h, int stride_w, int pad_h, int pad_w) {
-    
-    // Calculate output dimensions
-    int out_height = (input_h + 2 * pad_h - kernel_h) / stride_h + 1;
-    int out_width = (input_w + 2 * pad_w - kernel_w) / stride_w + 1;
-    
-    // Initialize output to zero
-    size_t output_size = batch_size * out_channels * out_height * out_width;
-    std::memset(output, 0, output_size * sizeof(float));
-    
-    for (int b = 0; b < batch_size; ++b) {
-        for (int oc = 0; oc < out_channels; ++oc) {
-            for (int oh = 0; oh < out_height; ++oh) {
-                for (int ow = 0; ow < out_width; ++ow) {
-                    
-                    // Calculate input region bounds
-                    int in_h_start = oh * stride_h - pad_h;
-                    int in_w_start = ow * stride_w - pad_w;
-                    
-                    float sum = 0.0f;
-                    
-                    for (int ic = 0; ic < in_channels; ++ic) {
-                        for (int kh = 0; kh < kernel_h; ++kh) {
-                            int in_h = in_h_start + kh;
-                            
-                            // Skip if outside input bounds
-                            if (in_h < 0 || in_h >= input_h) {
-                                continue;
-                            }
-                            
-                            int kw = 0;
-                            int in_w = in_w_start;
-                            
-                            // Skip negative width indices
-                            while (kw < kernel_w && in_w + kw < 0) {
-                                kw++;
-                            }
-                            
-                            size_t vl;
-                            for (; kw < kernel_w; kw += vl) {
-                                int remaining = kernel_w - kw;
-                                int valid_end = input_w - in_w - kw;
-                                int processable = std::min(remaining, valid_end);
-                                
-                                if (processable <= 0) break;
-                                
-                                vl = SET_VECTOR_LENGTH<float, M2>(processable);
-
-                                // Load input and kernel vectors using wrappers
-                                vfloat32m2_t v_input = VECTOR_LOAD<float, M2>(
-                                    &input[b * in_channels * input_h * input_w +
-                                           ic * input_h * input_w +
-                                           in_h * input_w + in_w + kw], vl);
-
-                                vfloat32m2_t v_kernel = VECTOR_LOAD<float, M2>(
-                                    &kernel[oc * in_channels * kernel_h * kernel_w +
-                                           ic * kernel_h * kernel_w +
-                                           kh * kernel_w + kw], vl);
-
-                                // Element-wise multiply
-                                vfloat32m2_t v_mult = VECTOR_MUL_VV<float, M2>(v_input, v_kernel, vl);
-
-                                // Reduce sum horizontally
-                                vfloat32m1_t v_zero = VECTOR_BROADCAST<float, M1>(0.0f, 1);
-                                vfloat32m1_t v_sum = __riscv_vfredusum_vs_f32m2_f32m1(v_mult, v_zero, vl);
-
-                                // Extract scalar sum and add to total
-                                sum += __riscv_vfmv_f_s_f32m1_f32(v_sum);
-                            }
-                        }
-                    }
-                    
-                    // Store the accumulated result
-                    output[b * out_channels * out_height * out_width +
-                           oc * out_height * out_width +
-                           oh * out_width + ow] = sum;
-                }
-            }
-        }
-    }
-}
-
-// RVV optimized 2D convolution (e32m4)
-void conv2d_e32m4(
-    const float* input, const float* kernel, float* output,
-    int batch_size, int in_channels, int out_channels,
-    int input_h, int input_w, int kernel_h, int kernel_w,
-    int stride_h, int stride_w, int pad_h, int pad_w) {
-    
-    // Calculate output dimensions
-    int out_height = (input_h + 2 * pad_h - kernel_h) / stride_h + 1;
-    int out_width = (input_w + 2 * pad_w - kernel_w) / stride_w + 1;
-    
-    // Initialize output to zero
-    size_t output_size = batch_size * out_channels * out_height * out_width;
-    std::memset(output, 0, output_size * sizeof(float));
-    
-    for (int b = 0; b < batch_size; ++b) {
-        for (int oc = 0; oc < out_channels; ++oc) {
-            for (int oh = 0; oh < out_height; ++oh) {
-                for (int ow = 0; ow < out_width; ++ow) {
-                    
-                    // Calculate input region bounds
-                    int in_h_start = oh * stride_h - pad_h;
-                    int in_w_start = ow * stride_w - pad_w;
-                    
-                    float sum = 0.0f;
-                    
-                    for (int ic = 0; ic < in_channels; ++ic) {
-                        for (int kh = 0; kh < kernel_h; ++kh) {
-                            int in_h = in_h_start + kh;
-                            
-                            // Skip if outside input bounds
-                            if (in_h < 0 || in_h >= input_h) {
-                                continue;
-                            }
-                            
-                            int kw = 0;
-                            int in_w = in_w_start;
-                            
-                            // Skip negative width indices
-                            while (kw < kernel_w && in_w + kw < 0) {
-                                kw++;
-                            }
-                            
-                            size_t vl;
-                            for (; kw < kernel_w; kw += vl) {
-                                int remaining = kernel_w - kw;
-                                int valid_end = input_w - in_w - kw;
-                                int processable = std::min(remaining, valid_end);
-                                
-                                if (processable <= 0) break;
-                                
-                                vl = SET_VECTOR_LENGTH<float, M4>(processable);
-
-                                // Load input and kernel vectors using wrappers
-                                vfloat32m4_t v_input = VECTOR_LOAD<float, M4>(
-                                    &input[b * in_channels * input_h * input_w +
-                                           ic * input_h * input_w +
-                                           in_h * input_w + in_w + kw], vl);
-
-                                vfloat32m4_t v_kernel = VECTOR_LOAD<float, M4>(
-                                    &kernel[oc * in_channels * kernel_h * kernel_w +
-                                           ic * kernel_h * kernel_w +
-                                           kh * kernel_w + kw], vl);
-
-                                // Element-wise multiply
-                                vfloat32m4_t v_mult = VECTOR_MUL_VV<float, M4>(v_input, v_kernel, vl);
-
-                                // Reduce sum horizontally
-                                vfloat32m1_t v_zero = VECTOR_BROADCAST<float, M1>(0.0f, 1);
-                                vfloat32m1_t v_sum = __riscv_vfredusum_vs_f32m4_f32m1(v_mult, v_zero, vl);
-
-                                // Extract scalar sum and add to total
-                                sum += __riscv_vfmv_f_s_f32m1_f32(v_sum);
-                            }
-                        }
-                    }
-                    
-                    // Store the accumulated result
-                    output[b * out_channels * out_height * out_width +
-                           oc * out_height * out_width +
-                           oh * out_width + ow] = sum;
-                }
-            }
-        }
-    }
-}
-
-// RVV optimized 2D convolution (e32m8)
-void conv2d_e32m8(
-    const float* input, const float* kernel, float* output,
-    int batch_size, int in_channels, int out_channels,
-    int input_h, int input_w, int kernel_h, int kernel_w,
-    int stride_h, int stride_w, int pad_h, int pad_w) {
-    
-    // Calculate output dimensions
-    int out_height = (input_h + 2 * pad_h - kernel_h) / stride_h + 1;
-    int out_width = (input_w + 2 * pad_w - kernel_w) / stride_w + 1;
-    
-    // Initialize output to zero
-    size_t output_size = batch_size * out_channels * out_height * out_width;
-    std::memset(output, 0, output_size * sizeof(float));
-    
-    for (int b = 0; b < batch_size; ++b) {
-        for (int oc = 0; oc < out_channels; ++oc) {
-            for (int oh = 0; oh < out_height; ++oh) {
-                for (int ow = 0; ow < out_width; ++ow) {
-                    
-                    // Calculate input region bounds
-                    int in_h_start = oh * stride_h - pad_h;
-                    int in_w_start = ow * stride_w - pad_w;
-                    
-                    float sum = 0.0f;
-                    
-                    for (int ic = 0; ic < in_channels; ++ic) {
-                        for (int kh = 0; kh < kernel_h; ++kh) {
-                            int in_h = in_h_start + kh;
-                            
-                            // Skip if outside input bounds
-                            if (in_h < 0 || in_h >= input_h) {
-                                continue;
-                            }
-                            
-                            int kw = 0;
-                            int in_w = in_w_start;
-                            
-                            // Skip negative width indices
-                            while (kw < kernel_w && in_w + kw < 0) {
-                                kw++;
-                            }
-                            
-                            size_t vl;
-                            for (; kw < kernel_w; kw += vl) {
-                                int remaining = kernel_w - kw;
-                                int valid_end = input_w - in_w - kw;
-                                int processable = std::min(remaining, valid_end);
-                                
-                                if (processable <= 0) break;
-                                
-                                vl = SET_VECTOR_LENGTH<float, M8>(processable);
-
-                                // Load input and kernel vectors using wrappers
-                                vfloat32m8_t v_input = VECTOR_LOAD<float, M8>(
-                                    &input[b * in_channels * input_h * input_w +
-                                           ic * input_h * input_w +
-                                           in_h * input_w + in_w + kw], vl);
-
-                                vfloat32m8_t v_kernel = VECTOR_LOAD<float, M8>(
-                                    &kernel[oc * in_channels * kernel_h * kernel_w +
-                                           ic * kernel_h * kernel_w +
-                                           kh * kernel_w + kw], vl);
-
-                                // Element-wise multiply
-                                vfloat32m8_t v_mult = VECTOR_MUL_VV<float, M8>(v_input, v_kernel, vl);
-
-                                // Reduce sum horizontally
-                                vfloat32m1_t v_zero = VECTOR_BROADCAST<float, M1>(0.0f, 1);
-                                vfloat32m1_t v_sum = __riscv_vfredusum_vs_f32m8_f32m1(v_mult, v_zero, vl);
-
-                                // Extract scalar sum and add to total
-                                sum += __riscv_vfmv_f_s_f32m1_f32(v_sum);
-                            }
-                        }
-                    }
-                    
-                    // Store the accumulated result
-                    output[b * out_channels * out_height * out_width +
-                           oc * out_height * out_width +
-                           oh * out_width + ow] = sum;
-                }
-            }
-        }
-    }
-}
-
-
-#ifdef BUILD_RVV_CONV_MAIN
-int main() {
-    std::cout << "=== RVV Conv2D Test ===" << std::endl;
-    
-    // Simple test parameters
-    const int batch_size = 1;
-    const int in_channels = 1;
-    const int out_channels = 1;
-    const int input_h = 4;
-    const int input_w = 4;
-    const int kernel_h = 3;
-    const int kernel_w = 3;
-    const int stride_h = 1;
-    const int stride_w = 1;
-    const int pad_h = 1;
-    const int pad_w = 1;
-    
-    // Calculate sizes
-    const int input_size = batch_size * in_channels * input_h * input_w;
-    const int kernel_size = out_channels * in_channels * kernel_h * kernel_w;
-    const int out_height = (input_h + 2 * pad_h - kernel_h) / stride_h + 1;
-    const int out_width = (input_w + 2 * pad_w - kernel_w) / stride_w + 1;
-    const int output_size = batch_size * out_channels * out_height * out_width;
-    
-    std::cout << "Input: " << batch_size << "x" << in_channels << "x" << input_h << "x" << input_w << std::endl;
-    std::cout << "Kernel: " << out_channels << "x" << in_channels << "x" << kernel_h << "x" << kernel_w << std::endl;
-    std::cout << "Output: " << batch_size << "x" << out_channels << "x" << out_height << "x" << out_width << std::endl;
-    
-    // Allocate memory
-    float* input = new float[input_size];
-    float* kernel = new float[kernel_size];
-    float* output_scalar = new float[output_size];
-    float* output_rvv = new float[output_size];
-    
-    // Initialize with simple values
-    for (int i = 0; i < input_size; ++i) {
-        input[i] = static_cast<float>(1.0f);
-    }
-    for (int i = 0; i < kernel_size; ++i) {
-        kernel[i] = static_cast<float>(1.0f) / 9.0f; // Average filter
-    }
-    
-    // Test scalar implementation
-    conv2d_scalar(input, kernel, output_scalar, batch_size, in_channels, out_channels,
-                  input_h, input_w, kernel_h, kernel_w, stride_h, stride_w, pad_h, pad_w);
-    
-    // Test RVV implementation (e32m1)
-    conv2d_e32m1(input, kernel, output_rvv, batch_size, in_channels, out_channels,
-                 input_h, input_w, kernel_h, kernel_w, stride_h, stride_w, pad_h, pad_w);
-    
-    // Compare outputs
-    bool correct = true;
-    const float tolerance = 1e-5f;
-    for (int i = 0; i < output_size; ++i) {
-        if (std::abs(output_scalar[i] - output_rvv[i]) > tolerance) {
-            std::cout << "Mismatch at index " << i << ": scalar=" << output_scalar[i] 
-                      << " rvv=" << output_rvv[i] << std::endl;
-            correct = false;
-            break;
-        }
-    }
-    
-    std::cout << "Correctness test: " << (correct ? "PASS" : "FAIL") << std::endl;
-    
-    // Print sample outputs
-    std::cout << "First 5 scalar outputs: ";
-    for (int i = 0; i < 5 && i < output_size; ++i) {
-        std::cout << output_scalar[i] << " ";
-    }
-    std::cout << std::endl;
-    
-    std::cout << "First 5 RVV outputs: ";
-    for (int i = 0; i < 5 && i < output_size; ++i) {
-        std::cout << output_rvv[i] << " ";
-    }
-    std::cout << std::endl;
-    
-    // Clean up
-    delete[] input;
-    delete[] kernel;
-    delete[] output_scalar;
-    delete[] output_rvv;
-    
-    std::cout << "Test completed!" << std::endl;
-    return 0;
-}
-#endif // BUILD_RVV_CONV_MAIN
